Cleanup
git-svn-id: svn://svn.code.sf.net/p/tass64/code/trunk@3222 111d7451-93fa-4e81-90d4-8039f76616e9

+++ unicode.c
@@ -657,7 +657,7 @@ void ctagsline_print(const uint8_t *l, FILE *f) {
             }
         } else i++;
         putc('\\', f);
-        putc((ch == '/' || ch == '\\') ? ch : '.', f);
+        putc((ch == '/' || ch == '\\') ? (int)ch : '.', f);
         len += 2;
         l = i;
     }
